fix: split interfaces into relevant folders
