+++ src/analyzer/index.ts
@@ -1,1 +1,2 @@
 export * from './analyzer';
+export * from './analyzer.interfaces';
