+++ src/client/index.ts
@@ -1,2 +1,3 @@
 export * from './superface-test';
+export * from './superface-test.interfaces';
 export * from './utils';
