+++ src/hash-generator/hash.interfaces.ts
