+++ src/hash-generator/index.ts
@@ -1,2 +1,3 @@
 export * from './generator';
+export * from './hash.interfaces';
 export * from './utils';
