+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from './client/superface-test';
-export * from './interfaces';
