--- src/interfaces/index.ts
@@ -1,6 +1,0 @@
-export * from './superface-test';
-export * from './hash';
-export * from './config';
-export * from './recording';
-export * from './matcher';
-export * from './analyzer';
