+++ src/matcher/index.ts
@@ -1,0 +1,4 @@
+export * from './matcher';
+export * from './matcher.interfaces';
+export * from './error-collector';
+export * from './utils';
