+++ src/recording/index.ts
@@ -2,3 +2,4 @@ export * from './recorder';
 export * from './controller';
 export * from './utils';
 export * from './replace';
+export * from './recording.interfaces';
