+++ src/recording/replace/index.ts
@@ -1,1 +1,2 @@
 export * from './replace';
+export * from './replace.interfaces';
