+++ src/superface/config/config.interfaces.ts
