+++ src/superface/config/index.ts
@@ -1,1 +1,2 @@
 export * from './prepare-superface';
+export * from './config.interfaces';
